rename all noun api to verb ones, fix readme typo
